Fixed toMorse somewhat
Added the XX for spaces, and X for delineating characters that don't have a space after. And not adding any X if it's at the end of the string. (unless the last char is a space, then add XX). It blows up if you give it too small of a buffer, so we need to figure out why.